feat: require Privy authentication for all protected pages
- Add reusable AuthGuard component for authentication protection
- Wrap all protected pages (chat, premiumchat, swap, transfer, bridge, overview, positions, echelon) with AuthGuard
- Ensure users must login with Privy before accessing any functionality
- Remove duplicate authentication checks from individual pages
- Improve code maintainability with centralized auth logic
