+++ frontend/app/bridge/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { usePrivy, WalletWithMetadata } from "@privy-io/react-auth";
-import { useRouter } from "next/navigation";
-import { useEffect, useState, useMemo, useRef } from "react";
+import { useState, useMemo, useRef } from "react";
 import { Sidebar } from "../components/sidebar";
 import { RightSidebar } from "../components/right-sidebar";
 import { ThemeToggle } from "../components/themeToggle";
+import { AuthGuard } from "../components/auth-guard";
 import { getTokenIconUrl } from "../utils/token-icons";
 import {
   Aptos,
@@ -44,8 +44,7 @@ const TOKENS = [
 ];
 
 export default function BridgePage() {
-  const { ready, authenticated, user } = usePrivy();
-  const router = useRouter();
+  const { authenticated, user } = usePrivy();
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
   const [isRightSidebarOpen, setIsRightSidebarOpen] = useState(false);
 
@@ -372,25 +371,8 @@ export default function BridgePage() {
     isValidEthereumAddress(recipientAddress) &&
     !bridging;
 
-  // Show loading while checking authentication status
-  if (!ready) {
-    return (
-      <div className="flex h-screen w-full items-center justify-center bg-zinc-50 font-sans dark:bg-black">
-        <div className="text-center">
-          <div className="text-lg text-zinc-600 dark:text-zinc-400">
-            Loading...
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Redirect if not authenticated
-  if (!authenticated) {
-    return null;
-  }
-
   return (
+    <AuthGuard>
     <div className="flex h-screen w-full overflow-hidden bg-zinc-50 dark:bg-black">
       <Sidebar isOpen={isSidebarOpen} onClose={() => setIsSidebarOpen(false)} />
 
@@ -942,5 +924,6 @@ export default function BridgePage() {
         onClose={() => setIsRightSidebarOpen(false)}
       />
     </div>
+    </AuthGuard>
   );
 }
