+++ frontend/app/positions/page.tsx
@@ -1,11 +1,12 @@
 "use client";
 
 import { usePrivy, WalletWithMetadata } from "@privy-io/react-auth";
-import { useRouter, useSearchParams } from "next/navigation";
+import { useSearchParams } from "next/navigation";
 import { useEffect, useState, useMemo, Suspense } from "react";
 import { Sidebar } from "../components/sidebar";
 import { RightSidebar } from "../components/right-sidebar";
 import { ThemeToggle } from "../components/themeToggle";
+import { AuthGuard } from "../components/auth-guard";
 import { SupplyModal } from "../components/supply-modal";
 import { BorrowModal } from "../components/borrow-modal";
 import { getTokenBySymbol, getVerifiedTokens } from "../utils/token-constants";
@@ -114,8 +115,7 @@ interface PortfolioResponse {
 }
 
 function PositionsPageContent() {
-  const { ready, authenticated, user } = usePrivy();
-  const router = useRouter();
+  const { authenticated, user } = usePrivy();
 
   const movementApiBase = getMovementApiBase();
 
@@ -245,12 +245,6 @@ function PositionsPageContent() {
   }, [movementApiBase]);
 
   // Redirect to home if not authenticated
-  useEffect(() => {
-    if (ready && !authenticated) {
-      router.push("/");
-    }
-  }, [ready, authenticated, router]);
-
   const marketPositions: MarketPosition[] = useMemo(() => {
     const verified = getVerifiedTokens();
     // Filter out legacy AptosCoin MOVE (use MOVE-FA instead which has higher utilization)
@@ -343,25 +337,8 @@ function PositionsPageContent() {
   const minRequiredEquityPercent =
     totalCollateral > 0 ? (minRequiredEquity / totalCollateral) * 100 : 0;
 
-  // Show loading while checking authentication status
-  if (!ready) {
-    return (
-      <div className="flex h-screen w-full items-center justify-center bg-zinc-50 font-sans dark:bg-black">
-        <div className="text-center">
-          <div className="text-lg text-zinc-600 dark:text-zinc-400">
-            Loading...
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Redirect if not authenticated
-  if (!authenticated) {
-    return null;
-  }
-
   return (
+    <AuthGuard>
     <div className="flex h-screen w-full overflow-hidden bg-zinc-50 dark:bg-black">
       <Sidebar isOpen={isSidebarOpen} onClose={() => setIsSidebarOpen(false)} />
 
@@ -928,6 +905,7 @@ function PositionsPageContent() {
         }}
       />
     </div>
+    </AuthGuard>
   );
 }
