+++ frontend/app/premiumchat/page.tsx
@@ -1,24 +1,23 @@
 "use client";
 
 import { usePrivy, WalletWithMetadata } from "@privy-io/react-auth";
-import { useRouter } from "next/navigation";
-import { useEffect, useState, useMemo } from "react";
+import { useState, useMemo } from "react";
 import { Sidebar } from "../components/sidebar";
 import { RightSidebar } from "../components/right-sidebar";
 import PremiumChat from "../components/chat/PremiumChat";
 import { ThemeToggle } from "../components/themeToggle";
+import { AuthGuard } from "../components/auth-guard";
 
 export default function PremiumChatPage() {
-  const { ready, authenticated, user } = usePrivy();
-  const router = useRouter();
+  const { authenticated, user } = usePrivy();
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
   const [isRightSidebarOpen, setIsRightSidebarOpen] = useState(false);
   const [selectedAgent, setSelectedAgent] = useState("premium_lending");
 
   // Get Movement wallet address (chainType is "aptos" for Movement wallets)
   const movementWallet = useMemo(() => {
-    // Only check for wallet when Privy is ready and user is authenticated
-    if (!ready || !authenticated || !user?.linkedAccounts) {
+    // Only check for wallet when user is authenticated
+    if (!authenticated || !user?.linkedAccounts) {
       return null;
     }
 
@@ -36,7 +35,7 @@ export default function PremiumChatPage() {
     ) as (WalletWithMetadata & { chainType?: string }) | undefined;
 
     return aptosWallet || null;
-  }, [user, ready, authenticated]);
+  }, [user, authenticated]);
 
   // Get the wallet address - ensure it's the full 66-character Movement/Aptos address
   const walletAddress = useMemo(() => {
@@ -50,32 +49,8 @@ export default function PremiumChatPage() {
     return null;
   }, [movementWallet]);
 
-  // Redirect to home if not authenticated
-  useEffect(() => {
-    if (ready && !authenticated) {
-      router.push("/");
-    }
-  }, [ready, authenticated, router]);
-
-  // Show loading while checking authentication status
-  if (!ready) {
-    return (
-      <div className="flex h-screen w-full items-center justify-center bg-zinc-50 font-sans dark:bg-black">
-        <div className="text-center">
-          <div className="text-lg text-zinc-600 dark:text-zinc-400">
-            Loading...
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Redirect if not authenticated (handled by useEffect, but show nothing while redirecting)
-  if (!authenticated) {
-    return null;
-  }
-
   return (
+    <AuthGuard>
     <div className="flex h-screen w-full overflow-hidden bg-zinc-50 dark:bg-black">
       <Sidebar
         isOpen={isSidebarOpen}
@@ -156,5 +131,6 @@ export default function PremiumChatPage() {
         onClose={() => setIsRightSidebarOpen(false)}
       />
     </div>
+    </AuthGuard>
   );
 }
